remove unnecssary functions
