+++ source/module_hamilt_lcao/hamilt_lcaodft/hamilt_lcao.cpp
@@ -342,7 +342,7 @@ HamiltLCAO<TK, TR>::HamiltLCAO(Gint_Gamma* GG_in,
         }
 #endif
         // TDDFT_velocity_gague
-        if (GlobalV::ESOLVER_TYPE == "tddft" && elecstate::H_TDDFT_pw::stype == 1)
+        if (TD_Velocity::tddft_velocity)
         {
             elecstate::H_TDDFT_pw::update_At();
             Operator<TK>* td_ekinetic = new TDEkinetic<OperatorLCAO<TK, TR>>(LM_in,
