+++ source/module_hamilt_lcao/hamilt_lcaodft/operator_lcao/td_ekinetic_lcao.cpp
@@ -110,7 +110,7 @@ void TDEkinetic<OperatorLCAO<TK, TR>>::calculate_HR()
                     std::complex<double>* tmp_c[3] = {nullptr, nullptr, nullptr};
                     for(int i = 0; i < 3; i++)
                     {
-                        tmp_c[i] = td_velocity.get_current_pointer()->get_current_term_pointer(i)->find_matrix(iat1, iat2, R_index2)->get_pointer();
+                        tmp_c[i] = td_velocity.get_current_term_pointer(i)->find_matrix(iat1, iat2, R_index2)->get_pointer();
                     }
                     this->cal_HR_IJR(iat1, iat2, paraV, dtau, tmp->get_pointer(), tmp_c, tmps->get_pointer());
                 }
@@ -349,8 +349,7 @@ void TDEkinetic<OperatorLCAO<TK, TR>>::contributeHR()
         //initialize current term if needed
         if (TD_Velocity::out_current)
         {
-            td_velocity.setup_current(this->ucell, this->Grid, paraV, this->intor_);
-            td_velocity.get_current_pointer()->initialize_current_term(this->hR_tmp, paraV);
+            td_velocity.initialize_current_term(this->hR_tmp, paraV);
         }
         // calculate the values in hR_tmp
         this->calculate_HR();
