--- source/module_hamilt_lcao/module_tddft/td_current.cpp
@@ -1,565 +1,0 @@
-#include "td_current.h"
-
-#include <unordered_map>
-#include "module_basis/module_ao/parallel_orbitals.h"
-#include "module_cell/module_neighbor/sltk_grid_driver.h"
-#include "module_cell/unitcell.h"
-#include "module_hamilt_lcao/module_hcontainer/hcontainer.h"
-#include "module_base/vector3.h"
-#include "module_hamilt_lcao/module_tddft/snap_psibeta_half_tddft.h"
-#include "module_base/timer.h"
-#ifdef _OPENMP
-#include <unordered_set>
-#endif
-
-#ifdef __LCAO
-
-bool TD_current::cal_grad = false;
-bool TD_current::cal_vcomm_r = false;
-
-TD_current::TD_current(const UnitCell* ucell_in,
-                       Grid_Driver* GridD_in,
-                       const Parallel_Orbitals* paraV,
-                       const ModuleBase::Vector3<double>& cart_At,
-                       const TwoCenterIntegrator* intor)
-    : ucell(ucell_in), Grid(GridD_in), paraV(paraV), cart_At(cart_At), intor_(intor)
-{   
-    if(cal_vcomm_r)
-    {
-        this->initialize_vcomm_r(GridD_in, paraV);
-    }
-    if(cal_grad)
-    {
-        this->initialize_grad_term(GridD_in, paraV);
-    }
-}
-TD_current::~TD_current()
-{
-    for (int dir=0;dir<3;dir++)
-    {
-        delete this->current_term[dir];
-    }
-}
-void TD_current::initialize_vcomm_r(Grid_Driver* GridD, const Parallel_Orbitals* paraV)
-{
-    ModuleBase::TITLE("TD_current", "initialize_vcomm_r");
-    ModuleBase::timer::tick("TD_current", "initialize_vcomm_r");
-    for (int dir=0;dir<3;dir++)
-    {
-        if (this->current_term[dir] == nullptr)
-        this->current_term[dir] = new hamilt::HContainer<std::complex<double>>(paraV);
-    }
-
-    this->adjs_vcommr.clear();
-    this->adjs_vcommr.reserve(this->ucell->nat);
-    for (int iat0 = 0; iat0 < ucell->nat; iat0++)
-    {
-        auto tau0 = ucell->get_tau(iat0);
-        int T0, I0;
-        ucell->iat2iait(iat0, &I0, &T0);
-        AdjacentAtomInfo adjs;
-        GridD->Find_atom(*ucell, tau0, T0, I0, &adjs);
-        std::vector<bool> is_adj(adjs.adj_num + 1, false);
-        for (int ad1 = 0; ad1 < adjs.adj_num + 1; ++ad1)
-        {
-            const int T1 = adjs.ntype[ad1];
-            const int I1 = adjs.natom[ad1];
-            const int iat1 = ucell->itia2iat(T1, I1);
-            const ModuleBase::Vector3<double>& tau1 = adjs.adjacent_tau[ad1];
-            const ModuleBase::Vector3<int>& R_index1 = adjs.box[ad1];
-            // choose the real adjacent atoms
-            const LCAO_Orbitals& orb = LCAO_Orbitals::get_const_instance();
-            // Note: the distance of atoms should less than the cutoff radius, 
-            // When equal, the theoretical value of matrix element is zero, 
-            // but the calculated value is not zero due to the numerical error, which would lead to result changes.
-            if (this->ucell->cal_dtau(iat0, iat1, R_index1).norm() * this->ucell->lat0
-                < orb.Phi[T1].getRcut() + this->ucell->infoNL.Beta[T0].get_rcut_max())
-            {
-                is_adj[ad1] = true;
-            }
-        }
-        filter_adjs(is_adj, adjs);
-        this->adjs_vcommr.push_back(adjs);
-        for (int ad1 = 0; ad1 < adjs.adj_num + 1; ++ad1)
-        {
-            const int T1 = adjs.ntype[ad1];
-            const int I1 = adjs.natom[ad1];
-            const int iat1 = ucell->itia2iat(T1, I1);
-            const ModuleBase::Vector3<int>& R_index1 = adjs.box[ad1];
-            for (int ad2 = 0; ad2 < adjs.adj_num + 1; ++ad2)
-            {
-                const int T2 = adjs.ntype[ad2];
-                const int I2 = adjs.natom[ad2];
-                const int iat2 = ucell->itia2iat(T2, I2);
-                ModuleBase::Vector3<int>& R_index2 = adjs.box[ad2];
-                if (paraV->get_col_size(iat2) <= 0 || paraV->get_row_size(iat1) <= 0)
-                {
-                    continue;
-                }
-                hamilt::AtomPair<std::complex<double>> tmp(iat1,
-                                         iat2,
-                                         R_index2.x - R_index1.x,
-                                         R_index2.y - R_index1.y,
-                                         R_index2.z - R_index1.z,
-                                         paraV);
-                for (int dir=0;dir<3;dir++)
-                {
-                    this->current_term[dir]->insert_pair(tmp);
-                }
-            }
-        }
-    }
-    // allocate the memory of BaseMatrix in cal_vcomm_r_IJR, and set the new values to zero
-    for (int dir=0;dir<3;dir++)
-    {
-        this->current_term[dir]->allocate(nullptr, true);
-    }
-    ModuleBase::timer::tick("TD_current", "initialize_vcomm_r");
-}
-void TD_current::initialize_grad_term(Grid_Driver* GridD, const Parallel_Orbitals* paraV)
-{
-    ModuleBase::TITLE("TD_current", "initialize_grad_term");
-    ModuleBase::timer::tick("TD_current", "initialize_grad_term");
-    for (int dir=0;dir<3;dir++)
-    {
-        if (this->current_term[dir] == nullptr)
-        this->current_term[dir] = new hamilt::HContainer<std::complex<double>>(paraV);
-    }
-
-    this->adjs_grad.clear();
-    this->adjs_grad.reserve(this->ucell->nat);
-    for (int iat1 = 0; iat1 < ucell->nat; iat1++)
-    {
-        auto tau1 = ucell->get_tau(iat1);
-        int T1, I1;
-        ucell->iat2iait(iat1, &I1, &T1);
-        AdjacentAtomInfo adjs;
-        GridD->Find_atom(*ucell, tau1, T1, I1, &adjs);
-        std::vector<bool> is_adj(adjs.adj_num + 1, false);
-        for (int ad1 = 0; ad1 < adjs.adj_num + 1; ++ad1)
-        {
-            const int T2 = adjs.ntype[ad1];
-            const int I2 = adjs.natom[ad1];
-            const int iat2 = ucell->itia2iat(T2, I2);
-            if (paraV->get_row_size(iat1) <= 0 || paraV->get_col_size(iat2) <= 0)
-            {
-                continue;
-            }
-            const ModuleBase::Vector3<int>& R_index2 = adjs.box[ad1];
-            // choose the real adjacent atoms
-            const LCAO_Orbitals& orb = LCAO_Orbitals::get_const_instance();
-            // Note: the distance of atoms should less than the cutoff radius, 
-            // When equal, the theoretical value of matrix element is zero, 
-            // but the calculated value is not zero due to the numerical error, which would lead to result changes.
-            if (this->ucell->cal_dtau(iat1, iat2, R_index2).norm() * this->ucell->lat0
-                < orb.Phi[T1].getRcut() + orb.Phi[T2].getRcut())
-            {
-                is_adj[ad1] = true;
-            }
-        }
-        filter_adjs(is_adj, adjs);
-        this->adjs_grad.push_back(adjs);
-        for (int ad = 0; ad < adjs.adj_num + 1; ++ad)
-        {
-            const int T2 = adjs.ntype[ad];
-            const int I2 = adjs.natom[ad];
-            int iat2 = ucell->itia2iat(T2, I2);
-            ModuleBase::Vector3<int>& R_index = adjs.box[ad];
-            hamilt::AtomPair<std::complex<double>> tmp(iat1, iat2, R_index.x, R_index.y, R_index.z, paraV);
-            for (int dir=0;dir<3;dir++)
-            {
-                this->current_term[dir]->insert_pair(tmp);
-            }
-        }
-    }
-    // allocate the memory of BaseMatrix in HR, and set the new values to zero
-    for (int dir=0;dir<3;dir++)
-    {
-        this->current_term[dir]->allocate(nullptr, true);
-    }
-
-    ModuleBase::timer::tick("EkineticNew", "initialize_HR");
-}
-void TD_current::initialize_current_term(const hamilt::HContainer<std::complex<double>>* HR, const Parallel_Orbitals* paraV)
-{
-    ModuleBase::TITLE("TD_current", "initialize_current_term");
-    ModuleBase::timer::tick("TD_current", "initialize_current_term");
-
-    for (int dir=0;dir<3;dir++)
-    {
-        if (this->current_term[dir] == nullptr)
-        this->current_term[dir] = new hamilt::HContainer<std::complex<double>>(paraV);
-    }
-
-    for (int i = 0; i < HR->size_atom_pairs(); ++i)
-    {
-        hamilt::AtomPair<std::complex<double>>& tmp = HR->get_atom_pair(i);
-        for (int ir = 0; ir < tmp.get_R_size(); ++ir)
-        {
-            const ModuleBase::Vector3<int> R_index = tmp.get_R_index(ir);
-            const int iat1 = tmp.get_atom_i();
-            const int iat2 = tmp.get_atom_j();
-
-            hamilt::AtomPair<std::complex<double>> tmp1(iat1, iat2, R_index, paraV);
-            for (int dir=0;dir<3;dir++)
-            {
-                this->current_term[dir]->insert_pair(tmp1);
-            }
-        }
-    }
-    for (int dir=0;dir<3;dir++)
-    {
-        this->current_term[dir]->allocate(nullptr, true);
-    }
-    ModuleBase::timer::tick("TDEkinetic", "initialize_HR_tmp");
-}
-
-void TD_current::calculate_vcomm_r()
-{
-    if(cal_vcomm_r == false)
-    {
-        return;
-    }
-    ModuleBase::TITLE("TDNonlocal", "calculate_HR");
-    ModuleBase::timer::tick("TDNonlocal", "calculate_HR");
-
-    const Parallel_Orbitals* paraV = this->current_term[0]->get_atom_pair(0).get_paraV();
-    const int npol = this->ucell->get_npol();
-    // 1. calculate <psi|beta> for each pair of atoms
-#ifdef _OPENMP
-#pragma omp parallel
-{
-    std::unordered_set<int> atom_row_list;
-    #pragma omp for
-    for (int iat0 = 0; iat0 < this->ucell->nat; iat0++)
-    {
-        atom_row_list.insert(iat0);
-    }
-#endif
-    for (int iat0 = 0; iat0 < this->ucell->nat; iat0++)
-    {
-        auto tau0 = ucell->get_tau(iat0);
-        int T0, I0;
-        ucell->iat2iait(iat0, &I0, &T0);
-        AdjacentAtomInfo& adjs = this->adjs_vcommr[iat0];
-
-        std::vector<std::vector<std::unordered_map<int, std::vector<std::complex<double>>>>> nlm_tot(4);
-        for(int dir=0;dir < 4;dir++)
-        {
-            nlm_tot[dir].resize(adjs.adj_num + 1);
-        }
-
-        for (int ad = 0; ad < adjs.adj_num + 1; ++ad)
-        {
-            const int T1 = adjs.ntype[ad];
-            const int I1 = adjs.natom[ad];
-            const int iat1 = ucell->itia2iat(T1, I1);
-            const ModuleBase::Vector3<double>& tau1 = adjs.adjacent_tau[ad];
-            const Atom* atom1 = &ucell->atoms[T1];
-
-            const LCAO_Orbitals& orb = LCAO_Orbitals::get_const_instance();
-            auto all_indexes = paraV->get_indexes_row(iat1);
-#ifdef _OPENMP
-            if(atom_row_list.find(iat1) == atom_row_list.end())
-            {
-                all_indexes.clear();
-            }
-#endif
-            auto col_indexes = paraV->get_indexes_col(iat1);
-            // insert col_indexes into all_indexes to get universal set with no repeat elements
-            all_indexes.insert(all_indexes.end(), col_indexes.begin(), col_indexes.end());
-            std::sort(all_indexes.begin(), all_indexes.end());
-            all_indexes.erase(std::unique(all_indexes.begin(), all_indexes.end()), all_indexes.end());
-            for (int iw1l = 0; iw1l < all_indexes.size(); iw1l += npol)
-            {
-                const int iw1 = all_indexes[iw1l] / npol;
-                std::vector<std::vector<std::complex<double>>> nlm;
-                // nlm is a vector of vectors, but size of outer vector is only 1 here
-                // If we are calculating force, we need also to store the gradient
-                // and size of outer vector is then 4
-                // inner loop : all projectors (L0,M0)
-                // snap_psibeta_half_tddft() are used to calculate <psi|exp(iAr)|beta> no matter whether to use new two-center or not for now.
-                module_tddft::snap_psibeta_half_tddft(orb,
-                                                     this->ucell->infoNL,
-                                                     nlm,
-                                                     tau1 * this->ucell->lat0,
-                                                     T1,
-                                                     atom1->iw2l[iw1],
-                                                     atom1->iw2m[iw1],
-                                                     atom1->iw2n[iw1],
-                                                     tau0 * this->ucell->lat0,
-                                                     T0,
-                                                     cart_At / 2.0,
-                                                     true);
-                for(int dir=0;dir<4;dir++)
-                {
-                    nlm_tot[dir][ad].insert({all_indexes[iw1l], nlm[dir]});
-                }
-            }
-        }
-// 2. calculate <psi_I|beta>D<beta|psi_{J,R}> for each pair of <IJR> atoms
-        for (int ad1 = 0; ad1 < adjs.adj_num + 1; ++ad1)
-        {
-            const int T1 = adjs.ntype[ad1];
-            const int I1 = adjs.natom[ad1];
-            const int iat1 = ucell->itia2iat(T1, I1);
-#ifdef _OPENMP
-            if(atom_row_list.find(iat1) == atom_row_list.end())
-            {
-                continue;
-            }
-#endif
-            ModuleBase::Vector3<int>& R_index1 = adjs.box[ad1];
-            for (int ad2 = 0; ad2 < adjs.adj_num + 1; ++ad2)
-            {
-                const int T2 = adjs.ntype[ad2];
-                const int I2 = adjs.natom[ad2];
-                const int iat2 = ucell->itia2iat(T2, I2);
-                ModuleBase::Vector3<int>& R_index2 = adjs.box[ad2];
-                ModuleBase::Vector3<int> R_vector(R_index2[0] - R_index1[0],
-                                                  R_index2[1] - R_index1[1],
-                                                  R_index2[2] - R_index1[2]);
-                hamilt::BaseMatrix<std::complex<double>>* tmpx = this->current_term[0]->find_matrix(iat1, iat2, R_vector[0], R_vector[1], R_vector[2]);
-                hamilt::BaseMatrix<std::complex<double>>* tmpy = this->current_term[1]->find_matrix(iat1, iat2, R_vector[0], R_vector[1], R_vector[2]);
-                hamilt::BaseMatrix<std::complex<double>>* tmpz = this->current_term[2]->find_matrix(iat1, iat2, R_vector[0], R_vector[1], R_vector[2]);
-                // if not found , skip this pair of atoms
-                if (tmpx != nullptr)
-                {
-                    this->cal_vcomm_r_IJR(iat1,
-                                     iat2,
-                                     T0,
-                                     paraV,
-                                     nlm_tot[0][ad1],
-                                     nlm_tot[1][ad1],
-                                     nlm_tot[2][ad1],
-                                     nlm_tot[3][ad1],
-                                     nlm_tot[0][ad2],
-                                     nlm_tot[1][ad2],
-                                     nlm_tot[2][ad2],
-                                     nlm_tot[3][ad2],
-                                     tmpx->get_pointer(),
-                                     tmpy->get_pointer(),
-                                     tmpz->get_pointer());
-                }
-            }
-        }
-    }
-#ifdef _OPENMP
-}
-#endif
-
-    ModuleBase::timer::tick("TDNonlocal", "calculate_HR");
-}
-
-// cal_HR_IJR()
-void TD_current::cal_vcomm_r_IJR(
-    const int& iat1,
-    const int& iat2,
-    const int& T0,
-    const Parallel_Orbitals* paraV,
-    const std::unordered_map<int, std::vector<std::complex<double>>>& nlm1_all,
-    const std::unordered_map<int, std::vector<std::complex<double>>>& nlm1x_all,
-    const std::unordered_map<int, std::vector<std::complex<double>>>& nlm1y_all,
-    const std::unordered_map<int, std::vector<std::complex<double>>>& nlm1z_all,
-    const std::unordered_map<int, std::vector<std::complex<double>>>& nlm2_all,
-    const std::unordered_map<int, std::vector<std::complex<double>>>& nlm2x_all,
-    const std::unordered_map<int, std::vector<std::complex<double>>>& nlm2y_all,
-    const std::unordered_map<int, std::vector<std::complex<double>>>& nlm2z_all,
-    std::complex<double>* data_pointerx,
-    std::complex<double>* data_pointery,
-    std::complex<double>* data_pointerz)
-{
-
-    // npol is the number of polarizations,
-    // 1 for non-magnetic (one Hamiltonian matrix only has spin-up or spin-down),
-    // 2 for magnetic (one Hamiltonian matrix has both spin-up and spin-down)
-    const int npol = this->ucell->get_npol();
-    // ---------------------------------------------
-    // calculate the Nonlocal matrix for each pair of orbitals
-    // ---------------------------------------------
-    double olm[3] = {0, 0, 0};
-    auto row_indexes = paraV->get_indexes_row(iat1);
-    auto col_indexes = paraV->get_indexes_col(iat2);
-    // step_trace = 0 for NSPIN=1,2; ={0, 1, local_col, local_col+1} for NSPIN=4
-    std::vector<int> step_trace(npol * npol, 0);
-    for (int is = 0; is < npol; is++)
-    {
-        for (int is2 = 0; is2 < npol; is2++)
-        {
-            //step_trace[is + is2 * npol] = col_indexes.size() * is + is2;
-            step_trace[is * npol + is2] = col_indexes.size() * is + is2;
-        }
-    }
-    // calculate the local matrix
-    const std::complex<double>* tmp_d = nullptr;
-    for (int iw1l = 0; iw1l < row_indexes.size(); iw1l += npol)
-    {
-        const std::vector<std::complex<double>>& nlm1 = nlm1_all.find(row_indexes[iw1l])->second;
-        const std::vector<std::complex<double>>& nlm1x = nlm1x_all.find(row_indexes[iw1l])->second;
-        const std::vector<std::complex<double>>& nlm1y = nlm1y_all.find(row_indexes[iw1l])->second;
-        const std::vector<std::complex<double>>& nlm1z = nlm1z_all.find(row_indexes[iw1l])->second;
-        for (int iw2l = 0; iw2l < col_indexes.size(); iw2l += npol)
-        {
-            const std::vector<std::complex<double>>& nlm2 = nlm2_all.find(col_indexes[iw2l])->second;
-            const std::vector<std::complex<double>>& nlm2x = nlm2x_all.find(col_indexes[iw2l])->second;
-            const std::vector<std::complex<double>>& nlm2y = nlm2y_all.find(col_indexes[iw2l])->second;
-            const std::vector<std::complex<double>>& nlm2z = nlm2z_all.find(col_indexes[iw2l])->second;
-#ifdef __DEBUG
-            assert(nlm1.size() == nlm2.size());
-#endif
-            for (int is = 0; is < npol * npol; ++is)
-            {
-                std::complex<double> nlm_tmpx = std::complex<double>{0,0};
-                std::complex<double> nlm_tmpy = std::complex<double>{0,0};
-                std::complex<double> nlm_tmpz = std::complex<double>{0,0};
-                std::complex<double> imag_unit = std::complex<double>{0,1};
-                for (int no = 0; no < this->ucell->atoms[T0].ncpp.non_zero_count_soc[is]; no++)
-                {
-                    const int p1 = this->ucell->atoms[T0].ncpp.index1_soc[is][no];
-                    const int p2 = this->ucell->atoms[T0].ncpp.index2_soc[is][no];
-                    //this->ucell->atoms[T0].ncpp.get_d(is, p2, p1, tmp_d);
-                    this->ucell->atoms[T0].ncpp.get_d(is, p1, p2, tmp_d);
-                    nlm_tmpx += (nlm1x[p1] * std::conj(nlm2[p2])-nlm1[p1] * std::conj(nlm2x[p2])) * (*tmp_d);
-                    nlm_tmpy += (nlm1y[p1] * std::conj(nlm2[p2])-nlm1[p1] * std::conj(nlm2y[p2])) * (*tmp_d);
-                    nlm_tmpz += (nlm1z[p1] * std::conj(nlm2[p2])-nlm1[p1] * std::conj(nlm2z[p2])) * (*tmp_d);
-                }
-                data_pointerx[step_trace[is]] -= imag_unit*nlm_tmpx/2.0;
-                data_pointery[step_trace[is]] -= imag_unit*nlm_tmpy/2.0;
-                data_pointerz[step_trace[is]] -= imag_unit*nlm_tmpz/2.0;
-            }
-            data_pointerx += npol;
-            data_pointery += npol;
-            data_pointerz += npol;
-        }
-        data_pointerx += (npol - 1) * col_indexes.size();
-        data_pointery += (npol - 1) * col_indexes.size();
-        data_pointerz += (npol - 1) * col_indexes.size();
-    }
-}
-void TD_current::calculate_grad_term()
-{
-    if(cal_grad == false)
-    {
-        return;
-    }
-    ModuleBase::TITLE("TDEkinetic", "calculate_HR");
-    if(this->current_term[0]==nullptr || this->current_term[0]->size_atom_pairs()<=0)
-    {
-        ModuleBase::WARNING_QUIT("TD_current::calculate_grad_term", "grad_term is nullptr or empty");
-    }
-    ModuleBase::timer::tick("TD_current", "calculate_grad_term");
-
-    const Parallel_Orbitals* paraV = this->current_term[0]->get_atom_pair(0).get_paraV();
-#ifdef _OPENMP
-#pragma omp parallel for
-#endif
-    for (int iat1 = 0; iat1 < this->ucell->nat; iat1++)
-    {
-        auto tau1 = ucell->get_tau(iat1);
-        int T1, I1;
-        ucell->iat2iait(iat1, &I1, &T1);
-        AdjacentAtomInfo& adjs = this->adjs_grad[iat1];
-        for (int ad = 0; ad < adjs.adj_num + 1; ++ad)
-        {
-            const int T2 = adjs.ntype[ad];
-            const int I2 = adjs.natom[ad];
-            const int iat2 = ucell->itia2iat(T2, I2);
-            const ModuleBase::Vector3<int>& R_index2 = adjs.box[ad];
-            ModuleBase::Vector3<double> dtau = this->ucell->cal_dtau(iat1, iat2, R_index2);
-
-            hamilt::BaseMatrix<std::complex<double>>* tmpx = this->current_term[0]->find_matrix(iat1, iat2, R_index2.x, R_index2.y, R_index2.z);
-            hamilt::BaseMatrix<std::complex<double>>* tmpy = this->current_term[1]->find_matrix(iat1, iat2, R_index2.x, R_index2.y, R_index2.z);
-            hamilt::BaseMatrix<std::complex<double>>* tmpz = this->current_term[2]->find_matrix(iat1, iat2, R_index2.x, R_index2.y, R_index2.z);
-            if (tmpx != nullptr)
-            {
-                this->cal_grad_IJR(iat1, iat2, paraV, dtau, tmpx->get_pointer(), tmpy->get_pointer(), tmpz->get_pointer());
-            }
-            else
-            {
-                ModuleBase::WARNING_QUIT("TD_current::calculate_grad_term", "R_index not found in HR");
-            }
-        }
-    }
-
-    ModuleBase::timer::tick("TD_current", "calculate_grad_term");
-}
-
-void TD_current::cal_grad_IJR(const int& iat1,
-                            const int& iat2,
-                            const Parallel_Orbitals* paraV,
-                            const ModuleBase::Vector3<double>& dtau,
-                            std::complex<double>* data_pointerx,
-                            std::complex<double>* data_pointery,
-                            std::complex<double>* data_pointerz)
-{
-    const LCAO_Orbitals& orb = LCAO_Orbitals::get_const_instance();
-    // ---------------------------------------------
-    // get info of orbitals of atom1 and atom2 from ucell
-    // ---------------------------------------------
-    int T1, I1;
-    this->ucell->iat2iait(iat1, &I1, &T1);
-    int T2, I2;
-    this->ucell->iat2iait(iat2, &I2, &T2);
-    Atom& atom1 = this->ucell->atoms[T1];
-    Atom& atom2 = this->ucell->atoms[T2];
-
-    // npol is the number of polarizations,
-    // 1 for non-magnetic (one Hamiltonian matrix only has spin-up or spin-down),
-    // 2 for magnetic (one Hamiltonian matrix has both spin-up and spin-down)
-    const int npol = this->ucell->get_npol();
-
-    const int* iw2l1 = atom1.iw2l;
-    const int* iw2n1 = atom1.iw2n;
-    const int* iw2m1 = atom1.iw2m;
-    const int* iw2l2 = atom2.iw2l;
-    const int* iw2n2 = atom2.iw2n;
-    const int* iw2m2 = atom2.iw2m;
-    
-    // ---------------------------------------------
-    // calculate the Ekinetic matrix for each pair of orbitals
-    // ---------------------------------------------
-    double grad[3] = {0, 0, 0};
-    auto row_indexes = paraV->get_indexes_row(iat1);
-    auto col_indexes = paraV->get_indexes_col(iat2);
-    const int step_trace = col_indexes.size() + 1;
-    for(int iw1l = 0; iw1l < row_indexes.size(); iw1l += npol)
-    {
-        const int iw1 = row_indexes[iw1l] / npol;
-        const int L1 = iw2l1[iw1];
-        const int N1 = iw2n1[iw1];
-        const int m1 = iw2m1[iw1];
-
-        // convert m (0,1,...2l) to M (-l, -l+1, ..., l-1, l)
-        int M1 = (m1 % 2 == 0) ? -m1/2 : (m1+1)/2;
-
-        for (int iw2l = 0; iw2l < col_indexes.size(); iw2l += npol)
-        {
-            const int iw2 = col_indexes[iw2l] / npol;
-            const int L2 = iw2l2[iw2];
-            const int N2 = iw2n2[iw2];
-            const int m2 = iw2m2[iw2];
-            
-            // convert m (0,1,...2l) to M (-l, -l+1, ..., l-1, l)
-            int M2 = (m2 % 2 == 0) ? -m2/2 : (m2+1)/2;
-
-            //calculate <psi|∇R|psi>, which equals to -<psi|∇r|psi>.
-            intor_->calculate(T1, L1, N1, M1,
-                    T2, L2, N2, M2, dtau * this->ucell->lat0, nullptr, grad);
-
-            for (int ipol = 0; ipol < npol; ipol++)
-            {
-                data_pointerx[ipol * step_trace]+= std::complex<double>(0,grad[0]);
-                data_pointery[ipol * step_trace]+= std::complex<double>(0,grad[1]);
-                data_pointerz[ipol * step_trace]+= std::complex<double>(0,grad[2]);
-            }
-            data_pointerx += npol;
-            data_pointery += npol;
-            data_pointerz += npol;
-        }
-        data_pointerx += (npol - 1) * col_indexes.size();
-        data_pointery += (npol - 1) * col_indexes.size();
-        data_pointerz += (npol - 1) * col_indexes.size();
-    }
-}
-
-#endif // __LCAO
